compat: [sinfl] 解决 llu  后缀字面量在 VS2010 下报错的问题 (#319)

+++ src/external/sinfl.h
@@ -414,7 +414,7 @@ sinfl_decompress(unsigned char *out, int cap, const unsigned char *in, int size)
 
       if ((unsigned short)len != (unsigned short)~nlen)
         return (int)(out-o);
-      if (len > (e - s.bitptr))
+      if ((int)len > (e - s.bitptr))
         return (int)(out-o);
 
       if (sinfl_unlikely(out + len > oe)) return -2;
@@ -542,8 +542,8 @@ sinfl_decompress(unsigned char *out, int cap, const unsigned char *in, int size)
           } else if (offs == 1) {
             /* rle match copying */
             unsigned int c = src[0];
-            unsigned int hw = (c << 24u) | (c << 16u) | (c << 8u) | (unsigned)c;
-            unsigned long long w = (unsigned long long)hw << 32llu | hw;
+            unsigned long long hw = (c << 24) | (c << 16) | (c << 8) | c;
+            unsigned long long w = (hw << 32) | hw;
             dst = sinfl_write64(dst, w);
             dst = sinfl_write64(dst, w);
             do dst = sinfl_write64(dst, w);
